Migrate from archlinux codespace to alpine-based codespace
The alpine image is much smaller than the archlinux image and allows me
to use an official vscode-dev-container base image. Using an official
image means I don't have to maintain the base image and all the bash
scripts that are necessary for configuring the container.

The downside is that some alpine packages are out of date. In particular,
the `go` package is one point release behind the latest version. I
mitigate this problem by manually installing the latest Go version. I'm
not using cgo, so musl libc should not create any problems.

I also added a few missing vscode settings and removed some recommended
extensions that should be optional to install. Finally, I changed the
CI image to an alpine image.