Fix default shell still not being zsh
Also, undo some of the docker hacks to keep the image small. I can't
think of any reason to optimize the image size. The Codespace behavior
is to build the image fresh when the new Codespace VM is created.
Optimizing the size only makes sense when the image will be pulled.