Remove CI status badge from README
* rename `Test` workflow to `PR`

There's no reason to run tests on `master` when the identical commit has
already been tested when it was a PR (pre-merge). Also, it should be
expected that `master` always passes tests. Since only rebase merging is
allowed, we can guarantee that the commit on `master` has been tested
by requiring that status checks pass in the repo settings on GitHub.